+++ MovieRating/MovieRating.Infrastructure.Data/MovieRating.Infrastructure.Data.csproj
@@ -4,4 +4,13 @@
     <TargetFramework>netstandard2.0</TargetFramework>
   </PropertyGroup>
 
+  <ItemGroup>
+    <PackageReference Include="Newtonsoft.Json" Version="11.0.2" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\MovieRating.Core.Entities\MovieRating.Core.Entities.csproj" />
+    <ProjectReference Include="..\MovieRating.Core\MovieRating.Core.csproj" />
+  </ItemGroup>
+
 </Project>
