+++ MovieRating/MovieRatingTest/MovieRatingServiceTest.cs
@@ -1,3 +1,4 @@
+using MovieRating.Infrastructure.Data;
 using System;
 using Xunit;
 
@@ -5,10 +6,18 @@ namespace MovieRatingTest
 {
     public class MovieRatingServiceTest
     {
-        [Fact]
-        public void Test1()
-        {
+        private MovieRatingRepository movieRR = new MovieRatingRepository(new JsonRead("TestRatings.json"));
 
+        //1
+        [Theory]
+        [InlineData(1, 5)]
+        [InlineData(2, 5)]
+        [InlineData(3, 5)]
+        [InlineData(4, 5)]
+        public void NumberReviewsFromReviewer(int input, int result)
+        {
+            int reviewsFromReviewer = movieRR.GetReviewsFromReviewer(input);
+            Assert.Equal(result, reviewsFromReviewer);
         }
     }
 }
