+++ MovieRating/MovieRatingTest/MovieRatingTest.csproj
@@ -12,4 +12,11 @@
     <PackageReference Include="xunit.runner.visualstudio" Version="2.3.1" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\MovieRating.Core.Entities\MovieRating.Core.Entities.csproj" />
+    <ProjectReference Include="..\MovieRating.Core\MovieRating.Core.csproj" />
+    <ProjectReference Include="..\MovieRating.Infrastructure.Data\MovieRating.Infrastructure.Data.csproj" />
+    <ProjectReference Include="..\MovieRating.UI.ConsoleApp\ConsoleApp.csproj" />
+  </ItemGroup>
+
 </Project>
